feat(type): types.ts move to ./shared/types and related changes
- tsconfig.json: add /shared/* to compile then, update config to use root cache
- *.types.ts,index.ts: move to shared/types
- package.json: update scripts to use root cache
- eslint: check shared/*.ts for lint types.ts
- eslint: モノレポ対応のため import/path 解決を共通化 & paths 設定追加
- tool-installer: use '@hared/types' instead of '@/shared/types'
