+++ .vscode/settings.json
@@ -1,0 +1,8 @@
+{
+  "editor.formatOnSave": true,
+  "json.format.enable": true,
+  "eslint.validate": [
+    "typescript",
+    "json"
+  ]
+}
