+++ packages/@aglabo-actions/tool-installer/shared/types/AgActionCommon.types.ts
