+++ packages/@aglabo-actions/tool-installer/shared/types/index.ts
