+++ packages/@aglabo-actions/tool-installer/src/helper/configs/__tests__/ToolConifgManager.spec.ts
@@ -10,13 +10,13 @@
 import { beforeEach, describe, expect, it } from 'vitest';
 
 // constants
-import { AgActionInstallerType } from '@/shared/types';
+import { AgActionInstallerType } from '@shared/types';
 
 // test units
 import { ToolConfigManager } from '../ToolConfigManager.class';
 
 // type
-import type { AgActionEgetToolConfig } from '@/shared/types';
+import type { AgActionEgetToolConfig } from '@shared/types';
 
 // test main
 describe('ToolConfigManager', () => {
