+++ packages/@aglabo-actions/tool-installer/src/helper/configs/__tests__/convertConfig.spec.ts
@@ -10,9 +10,9 @@
 import { describe, expect, it } from 'vitest';
 
 // Logger
-import { AgLogLevel } from '@/shared/types';
-import { AgActionInstallerType } from '@/shared/types';
 import { AgLogger } from '@/utils/logger/AgLogger.class';
+import { AgLogLevel } from '@shared/types';
+import { AgActionInstallerType } from '@shared/types';
 
 // test unit
 import { convertConfig } from '../convertConfig';
@@ -20,7 +20,7 @@ import { convertConfig } from '../convertConfig';
 import { parseJsoncConfig } from '../parseConfig';
 
 // types
-import type { AgActionEgetToolConfig } from '@/shared/types';
+import type { AgActionEgetToolConfig } from '@shared/types';
 
 // Log definition
 AgLogger.setLogLevel(AgLogLevel.DEBUG);
