+++ packages/@aglabo-actions/tool-installer/src/installer/AgActionInstallerExecutorsMap.ts
@@ -6,16 +6,15 @@
 // This software is released under the MIT License.
 // https://opensource.org/licenses/MIT
 
-// import
-
-import { AgActionInstallerType } from '@/shared/types';
+// typed module
+import { AgActionInstallerType } from '@shared/types';
 
 // executor
 import EgetInitializer from './executor/EgetInitializer';
 import NullExecutor from './executor/NullExecutor';
 
 // type
-import type { AgActionInstallerExecutorsMap } from '@/shared/types';
+import type { AgActionInstallerExecutorsMap } from '@shared/types';
 
 // map
 export const AgExecutorsMap: AgActionInstallerExecutorsMap = {
