+++ packages/@aglabo-actions/tool-installer/src/installer/executor/EgetInitializer.ts
@@ -22,7 +22,7 @@ import { AgActions_DEFAULT_TEMP_DIR, AgDir_WINGET_INSTALL_DIR } from '@shared/co
 // modules
 import { prepareInstallDirectory } from '@/utils/prepareInstallDirectory';
 
-import type { AgActionInstallerExecutor, AgActionInstallOptions } from '@/shared/types';
+import type { AgActionInstallerExecutor, AgActionInstallOptions } from '@shared/types';
 // types
 
 // routine
